+++ profiles/urls.py
@@ -2,7 +2,6 @@ from django.urls import path
 
 from .views import (
     ProfileView,
-    ReviewDetailView,
     UserCommentedReviewView,
     UserHostedMeetingView,
     UserHostedReviewView,
@@ -21,5 +20,4 @@ urlpatterns = [
         "commented/reviews/", UserCommentedReviewView.as_view(), name="commented_review"
     ),
     path("liked/reviews/", UserLikedReviewView.as_view(), name="liked_review"),
-    path("review/detail/<uuid:uuid>", ReviewDetailView.as_view(), name="review_detail"),
 ]
